renamed simulink libraries
The name "SEA_..." was quite inconsistent. The Toolbox name is "Compliant Joint Toolbox" and not "SEA Toolbox". So the prefix should be cjt. Similarly the Robotics Toolbox uses rtb_XYZ.
This might involve some annoying library link renaming, but should be acceptable.
The unit test for the library passes again. Some warnings to fix though.