all but one warning resolved
The remaining warning is this one:

Warning: When delay time is set to zero, the transport delay block 'simulinkBlockLibraryTest/JointMechanicalSubsystem/input_delay' is automatically set to support direct feedthrough. This may cause an algebraic loop. A Memory Block can be used in place of the Transport Delay to break the loop

It can be solved by a suitable mask code that replaces the input_delay block with a memory block in case the input delay mask parameter is set to zero.