On step towards resolving #50
The Simulink library all failed because of a bug in
A gain block holding the variable d_gl was introduced in 1046504d64bcf8f5d0c4971b7a345560b923156f but the variable d_gl was never declared in the block workspace. This caused all tests to fail due to a Simulink error.

Now all tests at least run and fail only by verification. But this only because we have not yet updated the checksums.  We'll do after resolving #48  and #49 .