removed Luenberger and Kalman observers
Due to #53 this commit removes the example files for the Luenberger and Kalman observer from the file tree in this branch.

The issue should be worked on in the branch:
https://github.com/geez0x1/CompliantJointToolbox/tree/dev_fix_issue_53

and merged after the issue is resolved.

In the library, both blocks are now commented out. Same for the corresponding unit test calls.